move fsst JNI bindings into converter code
