+++ converter/java/build.gradle
@@ -9,7 +9,7 @@ repositories {
     }
 }
 
-group 'com.covt'
+group 'com.mlt'
 version '1.0-SNAPSHOT'
 
 dependencies {
@@ -24,8 +24,6 @@ dependencies {
     implementation 'org.apache.orc:orc-core:1.8.1'
     implementation 'com.github.davidmoten:hilbert-curve:0.2.2'
     implementation 'org.roaringbitmap:RoaringBitmap:0.9.38'
-    implementation files('libs/fsst-1.0-SNAPSHOT.jar')
-
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
     testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
 }
@@ -43,3 +41,15 @@ java {
         languageVersion = JavaLanguageVersion.of(17)
     }
 }
+
+task compileWrapper(type: Exec) {
+    doFirst {
+        if (System.properties['os.name'].toLowerCase().contains('windows')) {
+            executable "./resources/compile.bat"
+        } else {
+            executable "./resources/compile.sh"
+        }
+    }
+}
+
+compileJava.dependsOn compileWrapper
