+++ converter/java/src/main/java/com/mlt/converter/encodings/fsst/SymbolTable.java
@@ -1,4 +1,4 @@
-package com.fsst;
+package com.mlt.converter.encodings.fsst;
 
 public record SymbolTable(byte[] symbols, int[] symbolLengths, byte[] compressedData){}
