--- encodings/fsst/cpp/README.md
@@ -1,1 +1,0 @@
-TODO: Link the [FSST](https://github.com/cwida/fsst) project and automate the build of the binaries.
