--- encodings/fsst/java/fsst-wrapper/.gitignore
@@ -1,3 +1,0 @@
-.gradle
-.idea
-build
